Alright here is a movement script for keyboards.
FINALLY, weve gone thru many freshman year reminders and now understand quaternions to a degree that I can edit them.. and embarisingly a bit of if else statements that should be changed into switch statements. ok now to move onto understanding the hand tracking sdk, which I belive I'll have a object attatched to the pointer finger and have the object "look at" and object on the same plane as the paddle. It will move only in the X and Z axis. blah ideas!