Got a simple tilt 5 input script up
It is now time for some input refining. Both the keyboard and the tilt 5 inputs are very basic, I would like to add some feel to them or add values to increase their usfullness.

some thoughts are having the axis of the tilt 5 stick bounce the quaternion over to where you point the stick. Maybe not imediatly, but hopefully snappy witha  bit of smoothing so the paddle is not too gittery. TIME FOR MORE RESEARCH AND MATHHHHH.